# [WIP] Perf-testing on hash functions