This is a WIP PR not intended to be merged.  It cherry-picks b00fbd8d8d9a572b3791dbf64127b79ec6981bbf from `lwaftr` to add the `snabbmark hash` function, extends the hash test to test variable-length keys, then adds the fnv hash.

Results:
```
$ for i in 1 2 4 6 8 12 16 24 32; do echo "./snabb snabbmark hash $i"; sudo ./snabb snabbmark hash $i; echo; done
./snabb snabbmark hash 1
No PMU available: single core cpu affinity required
baseline: 2.84 ns per iteration (result: 0)
jenkins hash: 18.28 ns per iteration (result: 236129928)
fnv hash: 4.42 ns per iteration (result: 2590564124)
murmur hash (32 bit): 17.93 ns per iteration (result: 1364076727)

./snabb snabbmark hash 2
No PMU available: single core cpu affinity required
baseline: 3.49 ns per iteration (result: 0)
jenkins hash: 24.06 ns per iteration (result: 2782625322)
fnv hash: 5.40 ns per iteration (result: 790051092)
murmur hash (32 bit): 18.25 ns per iteration (result: 821347078)

./snabb snabbmark hash 4
No PMU available: single core cpu affinity required
baseline: 3.16 ns per iteration (result: 0)
jenkins hash: 4.54 ns per iteration (result: 2355983250)
fnv hash: 6.34 ns per iteration (result: 3420582196)
murmur hash (32 bit): 8.22 ns per iteration (result: 593689054)

./snabb snabbmark hash 6
No PMU available: single core cpu affinity required
baseline: 3.21 ns per iteration (result: 0)
jenkins hash: 23.69 ns per iteration (result: 3787970914)
fnv hash: 7.11 ns per iteration (result: 3875976788)
murmur hash (32 bit): 18.18 ns per iteration (result: 4065731677)

./snabb snabbmark hash 8
No PMU available: single core cpu affinity required
baseline: 3.19 ns per iteration (result: 0)
jenkins hash: 7.49 ns per iteration (result: 236129928)
fnv hash: 8.16 ns per iteration (result: 3645689972)
murmur hash (32 bit): 10.14 ns per iteration (result: 1669671676)

./snabb snabbmark hash 12
No PMU available: single core cpu affinity required
baseline: 3.18 ns per iteration (result: 0)
jenkins hash: 103.57 ns per iteration (result: 3787970914)
fnv hash: 11.70 ns per iteration (result: 1767269300)
murmur hash (32 bit): 107.42 ns per iteration (result: 3644920907)

./snabb snabbmark hash 16
No PMU available: single core cpu affinity required
baseline: 3.20 ns per iteration (result: 0)
jenkins hash: 110.06 ns per iteration (result: 1799890284)
fnv hash: 14.04 ns per iteration (result: 2754996980)
murmur hash (32 bit): 114.29 ns per iteration (result: 2167721464)

./snabb snabbmark hash 24
No PMU available: single core cpu affinity required
baseline: 3.32 ns per iteration (result: 0)
jenkins hash: 119.25 ns per iteration (result: 2548393908)
fnv hash: 21.12 ns per iteration (result: 3617052020)
murmur hash (32 bit): 117.20 ns per iteration (result: 3708130561)

./snabb snabbmark hash 32
No PMU available: single core cpu affinity required
baseline: 2.71 ns per iteration (result: 0)
jenkins hash: 132.02 ns per iteration (result: 333401968)
fnv hash: 31.12 ns per iteration (result: 666669044)
murmur hash (32 bit): 129.31 ns per iteration (result: 407949976)

```

Conclusions: Although the pure-Lua fast cases for the Jenkins hash (for 4 and 8 bytes) perform very well, the generic version does not get specialized by LuaJIT to nice code.  I am not quite sure why but it does seem that the DynASM FNV hash is doing much better where we don't hit the Jenkins fast case.  Murmur is worse in all situations (from a speed POV).

Still it would seem that we can do better than the current FNV hash, especially considering that it's not unrolled and does byte-by-byte access.  It's worthy of a bit of code golf :)