@wingo 2017-05-26
Added a commit to unroll the hash function.  New results:

```
$ for i in 1 2 4 6 8 12 16 24 32; do echo "./snabb snabbmark hash $i"; sudo ./snabb snabbmark hash $i; echo; done
./snabb snabbmark hash 1
No PMU available: single core cpu affinity required
baseline: 2.84 ns per iteration (result: 0)
jenkins hash: 18.23 ns per iteration (result: 236129928)
fnv hash: 4.12 ns per iteration (result: 2590564124)
murmur hash (32 bit): 17.93 ns per iteration (result: 1364076727)

./snabb snabbmark hash 2
No PMU available: single core cpu affinity required
baseline: 3.49 ns per iteration (result: 0)
jenkins hash: 24.07 ns per iteration (result: 2782625322)
fnv hash: 4.78 ns per iteration (result: 790051092)
murmur hash (32 bit): 18.23 ns per iteration (result: 821347078)

./snabb snabbmark hash 4
No PMU available: single core cpu affinity required
baseline: 3.17 ns per iteration (result: 0)
jenkins hash: 4.42 ns per iteration (result: 2355983250)
fnv hash: 4.74 ns per iteration (result: 3420582196)
murmur hash (32 bit): 8.04 ns per iteration (result: 593689054)

./snabb snabbmark hash 6
No PMU available: single core cpu affinity required
baseline: 3.17 ns per iteration (result: 0)
jenkins hash: 23.76 ns per iteration (result: 3787970914)
fnv hash: 5.97 ns per iteration (result: 3875976788)
murmur hash (32 bit): 18.20 ns per iteration (result: 4065731677)

./snabb snabbmark hash 8
No PMU available: single core cpu affinity required
baseline: 3.18 ns per iteration (result: 0)
jenkins hash: 7.52 ns per iteration (result: 236129928)
fnv hash: 7.51 ns per iteration (result: 3645689972)
murmur hash (32 bit): 10.35 ns per iteration (result: 1669671676)

./snabb snabbmark hash 12
No PMU available: single core cpu affinity required
baseline: 3.23 ns per iteration (result: 0)
jenkins hash: 100.78 ns per iteration (result: 3787970914)
fnv hash: 10.37 ns per iteration (result: 1767269300)
murmur hash (32 bit): 105.52 ns per iteration (result: 3644920907)

./snabb snabbmark hash 16
No PMU available: single core cpu affinity required
baseline: 3.16 ns per iteration (result: 0)
jenkins hash: 102.95 ns per iteration (result: 1799890284)
fnv hash: 13.33 ns per iteration (result: 2754996980)
murmur hash (32 bit): 105.18 ns per iteration (result: 2167721464)

./snabb snabbmark hash 24
No PMU available: single core cpu affinity required
baseline: 3.26 ns per iteration (result: 0)
jenkins hash: 121.83 ns per iteration (result: 2548393908)
fnv hash: 20.30 ns per iteration (result: 3617052020)
murmur hash (32 bit): 123.42 ns per iteration (result: 3708130561)

./snabb snabbmark hash 32
No PMU available: single core cpu affinity required
baseline: 2.53 ns per iteration (result: 0)
jenkins hash: 127.32 ns per iteration (result: 333401968)
fnv hash: 29.89 ns per iteration (result: 666669044)
murmur hash (32 bit): 124.47 ns per iteration (result: 407949976)
```

Conclusion: Marginal improvement (I think).