@wingo 2017-05-26
OK!  I just hacked up a SipHash implementation, thinking that it might be faster than FNV as it reads 8 bytes at a time, even though its internal state is larger.  Also of course SipHash is a proper cryptographic hash function and it would be nice if we used one, for security reasons, if we can afford to pay for it anyway.

Results:
```
$ for i in 1 2 4 6 8 12 16 24 32; do echo "./snabb snabbmark hash $i"; sudo ./snabb snabbmark hash $i; echo; done
./snabb snabbmark hash 1
No PMU available: single core cpu affinity required
baseline: 2.85 ns per iteration (result: 0)
jenkins hash: 18.31 ns per iteration (result: 236129928)
fnv hash: 4.10 ns per iteration (result: 2590564124)
sip hash: 13.50 ns per iteration (result: 1962424773)
murmur hash (32 bit): 17.92 ns per iteration (result: 1364076727)

./snabb snabbmark hash 2
No PMU available: single core cpu affinity required
baseline: 3.50 ns per iteration (result: 0)
jenkins hash: 24.02 ns per iteration (result: 2782625322)
fnv hash: 4.72 ns per iteration (result: 790051092)
sip hash: 14.71 ns per iteration (result: 1361195817)
murmur hash (32 bit): 18.26 ns per iteration (result: 821347078)

./snabb snabbmark hash 4
No PMU available: single core cpu affinity required
baseline: 3.15 ns per iteration (result: 0)
jenkins hash: 4.42 ns per iteration (result: 2355983250)
fnv hash: 4.72 ns per iteration (result: 3420582196)
sip hash: 15.19 ns per iteration (result: 4025522647)
murmur hash (32 bit): 8.14 ns per iteration (result: 593689054)

./snabb snabbmark hash 6
No PMU available: single core cpu affinity required
baseline: 3.15 ns per iteration (result: 0)
jenkins hash: 23.66 ns per iteration (result: 3787970914)
fnv hash: 5.86 ns per iteration (result: 3875976788)
sip hash: 17.42 ns per iteration (result: 2868053420)
murmur hash (32 bit): 18.62 ns per iteration (result: 4065731677)

./snabb snabbmark hash 8
No PMU available: single core cpu affinity required
baseline: 3.16 ns per iteration (result: 0)
jenkins hash: 7.50 ns per iteration (result: 236129928)
fnv hash: 7.46 ns per iteration (result: 3645689972)
sip hash: 17.37 ns per iteration (result: 970163484)
murmur hash (32 bit): 10.24 ns per iteration (result: 1669671676)

./snabb snabbmark hash 12
No PMU available: single core cpu affinity required
baseline: 3.17 ns per iteration (result: 0)
jenkins hash: 104.29 ns per iteration (result: 3787970914)
fnv hash: 10.35 ns per iteration (result: 1767269300)
sip hash: 18.13 ns per iteration (result: 3449823072)
murmur hash (32 bit): 12.69 ns per iteration (result: 3644920907)

./snabb snabbmark hash 16
No PMU available: single core cpu affinity required
baseline: 3.16 ns per iteration (result: 0)
jenkins hash: 107.66 ns per iteration (result: 1799890284)
fnv hash: 13.28 ns per iteration (result: 2754996980)
sip hash: 21.56 ns per iteration (result: 2560676800)
murmur hash (32 bit): 111.03 ns per iteration (result: 2167721464)

./snabb snabbmark hash 24
No PMU available: single core cpu affinity required
baseline: 3.21 ns per iteration (result: 0)
jenkins hash: 117.11 ns per iteration (result: 2548393908)
fnv hash: 20.40 ns per iteration (result: 3617052020)
sip hash: 24.96 ns per iteration (result: 3846390737)
murmur hash (32 bit): 120.53 ns per iteration (result: 3708130561)

./snabb snabbmark hash 32
No PMU available: single core cpu affinity required
baseline: 2.53 ns per iteration (result: 0)
jenkins hash: 128.71 ns per iteration (result: 333401968)
fnv hash: 29.87 ns per iteration (result: 666669044)
sip hash: 31.88 ns per iteration (result: 2307970020)
murmur hash (32 bit): 128.18 ns per iteration (result: 407949976)
```

Summary: FNV hash has uniform cost of something just less than 1ns/byte, which allows it to scale really nicely to small keys.  However as keys get larger, SipHash gets more interesting, and approximately has parity with FNV at 32-byte keys.  Not sure what the conclusion for ctable should be though.