@wingo 2017-05-26
Reducing the number of rounds from 2,4 to 1,2 approximately halves SipHash cost.  Additionally since we are only hashing fixed-size keys, I think we can also tweak the handling of the tail word.  Looks like to me SipHash1,2 will be a winner.